Remove Friend feature, rename packages, tests running
