--- db/4_create_friendship.json
@@ -1,50 +1,0 @@
-{
-  "databaseChangeLog": [
-    {
-      "changeSet": {
-        "id": "4",
-        "author": "zach",
-        "changes": [
-          {
-            "createTable": {
-              "tableName": "friendship",
-              "columns": [
-                {
-                  "column": {
-                    "name": "user1_id",
-                    "type": "int",
-                    "constraints": {
-                      "primaryKey": true,
-                      "primaryKeyName": "users_pk",
-                      "nullable": false,
-                      "references": "public.user(id)",
-                      "foreignKeyName": "user1_id_fk"
-                    }
-                  }
-                },
-                {
-                  "column": {
-                    "name": "user2_id",
-                    "type": "int",
-                    "constraints": {
-                      "primaryKey": true,
-                      "primaryKeyName": "users_pk",
-                      "nullable": false,
-                      "references": "public.user(id)",
-                      "foreignKeyName": "user2_id_fk"
-                    }
-                  }
-                }
-              ]
-            }
-          },
-          {
-            "sql": {
-              "sql": "ALTER TABLE public.friendship ADD CONSTRAINT user1_lessthan_user2_check CHECK (user1_id < user2_id);"
-            }
-          }
-        ]
-      }
-    }
-  ]
-}
