--- db/5_create_friend_request.json
@@ -1,45 +1,0 @@
-{
-  "databaseChangeLog": [
-    {
-      "changeSet": {
-        "id": "5",
-        "author": "zach",
-        "changes": [
-          {
-            "createTable": {
-              "tableName": "friend_request",
-              "columns": [
-                {
-                  "column": {
-                    "name": "requester_id",
-                    "type": "int",
-                    "constraints": {
-                      "primaryKey": true,
-                      "primaryKeyName": "req_pk",
-                      "nullable": false,
-                      "references": "public.user(id)",
-                      "foreignKeyName": "requester_id_fk"
-                    }
-                  }
-                },
-                {
-                  "column": {
-                    "name": "requested_id",
-                    "type": "int",
-                    "constraints": {
-                      "primaryKey": true,
-                      "primaryKeyName": "req_pk",
-                      "nullable": false,
-                      "references": "public.user(id)",
-                      "foreignKeyName": "requested_id_fk"
-                    }
-                  }
-                }
-              ]
-            }
-          }
-        ]
-      }
-    }
-  ]
-}
