--- src/main/kotlin/sh/zachwal/button/db/jdbi/FriendRequest.kt
@@ -1,6 +1,0 @@
-package sh.zachwal.authserver.db.jdbi
-
-data class FriendRequest(
-    val requester: User,
-    val requested: User
-)
