--- src/main/kotlin/sh/zachwal/button/friends/api/AcceptFriendRequest.kt
@@ -1,5 +1,0 @@
-package sh.zachwal.button.friends.api
-
-data class AcceptFriendRequest(
-    val fromUserId: Long
-)
