--- src/main/kotlin/sh/zachwal/button/friends/api/SendFriendRequest.kt
@@ -1,5 +1,0 @@
-package sh.zachwal.button.friends.api
-
-data class SendFriendRequest(
-    val toUserId: Long
-)
