+++ src/main/kotlin/sh/zachwal/button/guice/JdbiModule.kt
@@ -7,11 +7,9 @@ import org.jdbi.v3.core.kotlin.KotlinPlugin
 import org.jdbi.v3.postgres.PostgresPlugin
 import org.jdbi.v3.sqlobject.kotlin.KotlinSqlObjectPlugin
 import org.jdbi.v3.sqlobject.kotlin.onDemand
-import sh.zachwal.authserver.db.dao.FriendDAO
-import sh.zachwal.authserver.db.dao.FriendRequestDAO
-import sh.zachwal.authserver.db.dao.SessionDAO
-import sh.zachwal.authserver.db.dao.UserDAO
-import sh.zachwal.authserver.db.dao.UserRoleDAO
+import sh.zachwal.button.db.dao.SessionDAO
+import sh.zachwal.button.db.dao.UserDAO
+import sh.zachwal.button.db.dao.UserRoleDAO
 import javax.sql.DataSource
 
 class JdbiModule : AbstractModule() {
@@ -31,10 +29,4 @@ class JdbiModule : AbstractModule() {
 
     @Provides
     fun roleDao(jdbi: Jdbi): UserRoleDAO = jdbi.onDemand()
-
-    @Provides
-    fun friendDao(jdbi: Jdbi): FriendDAO = jdbi.onDemand()
-
-    @Provides
-    fun friendRequestDAO(jdbi: Jdbi): FriendRequestDAO = jdbi.onDemand()
 }
